requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-core
python-dotenv

